Add tests and fix types for formatElapsedTime
Added a test suite for the `formatElapsedTime` function in `utils.cpp` to ensure proper time formatting. Also updated the formatting in `utils.cpp` to cast `uint32_t` variables to `unsigned long` when using the `%lu` specifier to prevent compiler warnings across different host/target architectures.

+++ utils.cpp
@@ -696,8 +696,8 @@ void formatElapsedTime(char* timeStr, uint32_t timeVal, bool noDays) {
   secs = secs - (mins * 60); //subtract the converted seconds to minutes in order to display 59 secs max
   mins = mins - (hours * 60); //subtract the converted minutes to hours in order to display 59 minutes max
   hours = hours - (days * 24); //subtract the converted hours to days in order to display 23 hours max
-  if (noDays) sprintf(timeStr, "%02lu:%02lu:%02lu", hours, mins, secs);
-  else sprintf(timeStr, "%lu-%02lu:%02lu:%02lu", days, hours, mins, secs);
+  if (noDays) sprintf(timeStr, "%02lu:%02lu:%02lu", (unsigned long)hours, (unsigned long)mins, (unsigned long)secs);
+  else sprintf(timeStr, "%lu-%02lu:%02lu:%02lu", (unsigned long)days, (unsigned long)hours, (unsigned long)mins, (unsigned long)secs);
 }
 
 static time_t setAlarm(uint8_t alarmHour) {
